+++ src/components/dialog/dialog.ts
@@ -3,11 +3,11 @@ import { customElement, property, query, state } from 'lit/decorators';
 import { classMap } from 'lit-html/directives/class-map';
 import { ifDefined } from 'lit-html/directives/if-defined';
 import { event, EventEmitter, watch } from '../../internal/decorators';
-import styles from 'sass:./dialog.scss';
 import { lockBodyScrolling, unlockBodyScrolling } from '../../internal/scroll';
 import { hasSlot } from '../../internal/slot';
 import { isPreventScrollSupported } from '../../internal/support';
 import Modal from '../../internal/modal';
+import styles from 'sass:./dialog.scss';
 
 const hasPreventScroll = isPreventScrollSupported();
 
@@ -41,6 +41,7 @@ export default class SlDialog extends LitElement {
 
   private componentId = `dialog-${++id}`;
   private modal: Modal;
+  private originalTrigger: HTMLElement | null;
   private willShow = false;
   private willHide = false;
 
@@ -86,10 +87,7 @@ export default class SlDialog extends LitElement {
   connectedCallback() {
     super.connectedCallback();
 
-    this.modal = new Modal(this, {
-      onfocusOut: () => this.panel.focus()
-    });
-
+    this.modal = new Modal(this);
     this.handleSlotChange();
 
     // Show on init if open
@@ -115,6 +113,7 @@ export default class SlDialog extends LitElement {
       return;
     }
 
+    this.originalTrigger = document.activeElement as HTMLElement;
     this.willShow = true;
     this.isVisible = true;
     this.open = true;
@@ -169,6 +168,12 @@ export default class SlDialog extends LitElement {
     this.open = false;
     this.modal.deactivate();
 
+    // Restore focus to the original trigger
+    const trigger = this.originalTrigger;
+    if (trigger && typeof trigger.focus === 'function') {
+      setTimeout(() => trigger.focus());
+    }
+
     unlockBodyScrolling(this);
   }
 
@@ -178,6 +183,7 @@ export default class SlDialog extends LitElement {
 
   handleKeyDown(event: KeyboardEvent) {
     if (event.key === 'Escape') {
+      event.stopPropagation();
       this.hide();
     }
   }